+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [Unreleased]
 **Adicionados**
-* Método upwind explícito. O método upwind é generalizado com o método central em uma função `_explicit()`, diferenciados por uma constante `𝛉`.
+* `_set_𝛉()`, determinar o valor de `𝛉`, que difere os métodos explícitos e implícitos.
+* `_explicit()`, métodos de diferenças finitas centrais / upwind explícitos.
+* `_implicit()`, métodos de diferenças finitas centrais / upwind implícitos.
 * `__init__()`, cria o atributo `methods`, uma lista com os métodos numéricos implementados.
 * `_check_arguments()`, `_check_tuple()`, `_check_len()` e `_check_mthd()` para verificações dos argumentos de `solve()`.